refactor: extract input mouse dispatcher
